fix/broken-plugin
- Created since duration was just 1ms. Obsisdian provides no guarantee
the plugin event will fire that quickly. In testing, 7ms was the
average. I've extended this to 5 seconds, as I can't see any reason to
need it to be more recent than that.

- The plugin relied upon the header having an empty line after it in the
template. Where this was not the case, the TODOs would not be injected.
I've removed that requirement.

- Where no header is selected in settings, we should add a new line
before injecting the TODOs, instead of appending to whatever is
currently on the bottom line.