Fix how it checks if Daily Notes are enabled
It seems that the Obsidian API for reading configuration has changed. In
the API, there does not seem to be any configuration object available.
Instead, there is a configDir property exposed on the Vault object.

https://github.com/obsidianmd/obsidian-api/blob/b0aa06eab03d8c39c0733949da5aa1edc2084b01/obsidian.d.ts#L2936-L2941

Although I could not find any reference to it, it seems that Obsidian
stores the list of the active core plugins under
${configDir}/core-plugins.json.

Obsidian API reference:
https://github.com/obsidianmd/obsidian-api/blob/master/obsidian.d.ts

This fixes https://github.com/shichongrui/obsidian-rollover-daily-todos/issues/17